+++ mojo/helpers/crypto/__init__.py
@@ -2,3 +2,4 @@ from .aes import encrypt, decrypt, decrypt_ecb
 from .utils import random_bytes, random_string, b64_encode, b64_decode
 from .hash import hash
 from .sign import generate_signature as sign, verify_signature as verify, sign_for_group
+from .keys import secret_keys